Bringing code kicking and screaming to the MVC age
A bunch of changes and reorganising code, so that everything works in
MVC fashion and responsibilities are put to a bit more logical classes.
The code is ugly (stuff got thrown to new locations) and language name
loading is as of yet broken. And, of course, not everything has been tested.

darcs-hash:20080212163931-36ee1-296b68b4a318df8c5998b125c11c732510810cd1.gz