More work on word classes
The data model is starting to shape up. Now has the appropriate schema
(and the XSD is clearer), and the test proggy does exactly what I
expect it to do. Some parts of the editor coded, but it crashes, because
it's not exactly well developed yet and the list model is crappy. Yay.