Messed around with UI
I've almost succeeded in making this stuff actually work. Everything
works now EXCEPT the vertical expansion of content, and I have no fucking
idea why it does that.

The upcoming plan is to rebuild this bastard using UI toolkit that
would make a little bit more sense. Like JavaFX or something.
Unfortunately, Oracle JUST RANDOMLY HAPPENED to withdraw Scene Builder
from current version of JDK and building it from source is the only
option and it needs Gradle and Mercurial and other bullshit which
I don't have. So all of this has to wait.